requirements.txt
streamlit
requests
python-dotenv
sentence-transformers
scikit-learn